Initial Commit
This is the initial commit for the project. The first iteration here
borrows heavily from the Go and JS implementations. Things aren't
quite working yet, but it's at least compiling. It's still missing the
meat of the protocol, this just starts building some of the helpers for
writing data in a subunit stream. The next step is to build out the glue
building a stream from events.