Add .travis.yml to repo
While this repo doesn't do anything of real value yet, or have any
testing to speak of I think it's important to verify that we can at
least build the code. This is my first real project in rust and
verifying that everything at least compiles has been a struggle given
the scrutiny of the compiler. So having a CI system verify whatever is
changed as the library grows is useful.