Flesh out the actual writing in Event write() method
This commit fleshes out the actual writing in the Event write() method.
While we're still missing support for building a few more pieces of a
packet this adds the code to take the pieces we have build the crc32
checksum and write everything in the proper order out the provided
writer. As we add the missing packet components this also gives us the
framework for doing this.