Finish pieces of packet writing function
This commit adds the 3 missing components of an event packet: mime
type, file_content, and routing code for the write method. This
conceptually concludes all the pieces necessary to write out a subunit
packet for an event. While I doubt this works at all, since there is no
testing on anything this at least has all the pieces now.