Fixes for test event packets
This commit implements a number of fixes for generating test_event
packets. A test was added to verify we at least don't error while
generating a packet from an event. There is no verification of the
actual subunit data in the test yet, that'll be a future step. Using
that test and manually passing the subunit data to existing subunit
utilities (like subunit-trace and subunit-2to1) the data was able to be
parsed properly.