Expand and cleanup existing tests
This commit adds a second test case to test a full test event with a
file attachment. This involves writing 2 packets to the buffer, one for
the test start and the other for when it finishes. At the same time a
comment was added to the tests which can be used to write the generated
streams from the tests to a file in /tmp. This was how I've been
manually verifying the contents of generated packets using existing
subunit utilities.

Additionally, the boilerplate it works test is removed since we don't
need it anymore.