Add subunit reading support
This commit adds an implementation for reading subunit files. The
parse_subunit() function is added to take any object implementing
the std::io::Read trait parse it's contents as subunit data and return
a vector of event structures for each subunit packet. To verify it's
functionality the write tests are modified to leverage this function
to compare the input events with the processed output.