Make Event struct fields public
The Event struct is the primary entrypoint for writing subunit packets
and if an external lib or tool wants to create a packet they need to be
able to create Event structs for packets. The problem is that by default
fields are set to private so you can't set them when instantiating the
struct externally. This commit fixes this by setting all the fields in
Event to public. The alternative pattern is to create a method to
instantiate the struct, but because of the number of fields and that
they are all optional just setting the fields public is a simpler
approach. Especially when you look at the usage pattern of Event is to
either set it once to write, or access it directly to read subunit.