Don't loop over buffers to write the contents
There are a couple stages where subunit content is buffered in the
writing subunit path. When the subunit data needs to move between these
buffers we were previously looping over the buffer and writing one byte
at a time into the new buffer which was slow and inefficient. It was
also completely unnecessary because we can just pass the buffer directly
t the write() method being used. This commit updates all the instances
where we were looping over the buffer with just a direct write() call.
This significantly improves the performance of writing packets,
especially large packets.