Fix warnings on write() calls
The previous commit switched to calling the write() methods directly
when writing the full packet to a buffer. However, that commit neglected
to handle the Result return from the write() method. This led to a
compiler warning for each one of those new lines. This commit fixes that
by handling the Result properly.