Fix issues reported by cargo clippy
This commit goes through all the errors and warnings (since we have
`#![deny(warnings)]` warnings set in the src and cleans them up per the
suggestions of the tool. The only exception is the unreadable_literal,
mostly due to my personal preference of thinking that adding '_' to
numbers makes it more unreadable. That failure mode is added to the
allow list.