Switch crc lib to crc32fast
This commit updates the crc library we use to be crc32fast, we only are
concerned with generating a crc32 hash and the crc library api makes
this a bit more involved than it needs to be in newer versions. The
crc32fast library enables us to easily generate this hash and do it
faster by levarging SIMD instrinsics.